# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for ML Master Data",
        "title": "ML Master Data API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/",
    "paths": {
...
        "/tournaments": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get all tournaments",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tournament"
                ],
                "summary": "Get all tournaments",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Tournament"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new tournament with the given name",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tournament"
                ],
                "summary": "Create a new tournament",
                "parameters": [
                    {
                        "description": "Tournament request",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TournamentRequestDto"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Tournament"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tournaments/{tournamentID}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get a tournament by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tournament"
                ],
                "summary": "Get a tournament by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tournament ID",
                        "name": "tournamentID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Tournament"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Tournament not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update a tournament with the given name",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tournament"
                ],
                "summary": "Update a tournament",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tournament ID",
                        "name": "tournamentID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Tournament request",
                        "name": "dto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TournamentRequestDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Tournament"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Tournament not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete a tournament with the given tournament ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tournament"
                ],
                "summary": "Delete a tournament",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tournament ID",
                        "name": "tournamentID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tournament deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Tournament not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.TournamentRequestDto": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "models.Tournament": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "tournament_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}